+++ Source/Krypton Components/Krypton.Themes/Palette Builtin/Office 2010/Extra Themes/PaletteOffice2010BlueLightMode.cs
@@ -5116,7 +5116,7 @@ public class KryptonColorTable2010BlueLightMode : KryptonColorTable
     /// <summary>
     /// Gets the text color used on the tool strip.
     /// </summary>
-    public override Color ToolStripText => Colors.Resolve(SchemeBaseColors.ToolStripText, SchemeBaseColors.StatusStripText);
+    public override Color ToolStripText => Colors.ResolveToolStripText(SchemeBaseColors.StatusStripText);
 
     #endregion