+++ Source/Krypton Components/Krypton.Toolkit/Palette Controls/KryptonColorTableVisualStudio.cs
@@ -493,7 +493,7 @@ public class KryptonColorTableVisualStudio : KryptonColorTable
     /// <summary>
     /// Gets the text color used on the tool strip.
     /// </summary>
-    public override Color ToolStripText => _colors.Resolve(SchemeBaseColors.ToolStripText, SchemeBaseColors.StatusStripText);
+    public override Color ToolStripText => _colors.ResolveToolStripText(SchemeBaseColors.StatusStripText);
 
     #endregion