+++ Source/Krypton Components/Krypton.Toolkit/Palette Controls/KryptonVisualStudio2010With2010ColorTable.cs
@@ -487,7 +487,7 @@ public class KryptonVisualStudio2010With2010ColorTable : KryptonColorTable
     /// <summary>
     /// Gets the text color used on the tool strip.
     /// </summary>
-    public override Color ToolStripText => Colors.Resolve(SchemeBaseColors.ToolStripText, SchemeBaseColors.StatusStripText);
+    public override Color ToolStripText => Colors.ResolveToolStripText(SchemeBaseColors.StatusStripText);
 
     #endregion